fix: sticker interaction
