+++ src/pages/sticker-vote/Finalists/data.json
