+++ src/pages/sticker-vote/Header/index.tsx
@@ -2,9 +2,12 @@ import { makeStyles } from "tss-react/mui"
 
 import { Box, Container, Stack, SvgIcon, Typography } from "@mui/material"
 
-import { ReactComponent as TadaSvg } from "@/assets/svgs/sticker-contest-result/tada.svg"
+import { ReactComponent as CalenderSvg } from "@/assets/svgs/sticker-contest/calender.svg"
+import Button from "@/components/Button"
+import { STICKER_CONTEST_DISCORD_VOTE_URL } from "@/constants"
 import useCheckViewport from "@/hooks/useCheckViewport"
-import GridBackground from "@/pages/sticker-contest/Header/GridBackground"
+
+import GridBackground from "../components/GridBackground"
 
 const useStyles = makeStyles()(theme => ({
   announcement: {
@@ -16,9 +19,13 @@ const useStyles = makeStyles()(theme => ({
     borderRadius: "1rem",
     backgroundColor: "#FFE1DC",
     padding: "0.7rem 2rem",
+    [theme.breakpoints.down("md")]: {
+      top: "-4rem",
+    },
     [theme.breakpoints.down("sm")]: {
       padding: "0.6rem 2rem",
       borderWidth: "2px",
+      top: "-3rem",
     },
   },
 
@@ -34,7 +41,6 @@ const useStyles = makeStyles()(theme => ({
   award: {
     padding: "0.8rem 2rem",
     border: `3px solid ${theme.palette.text.primary}`,
-    height: "8rem",
     backgroundColor: theme.palette.themeBackground.light,
     borderRadius: "1.6rem",
     textAlign: "center",
@@ -44,16 +50,24 @@ const useStyles = makeStyles()(theme => ({
   },
 }))
 
-const StickerContestResult = () => {
+const StickerContest = () => {
+  const { classes } = useStyles()
+  const { isMobile } = useCheckViewport()
+
   return (
-    <Box sx={{ backgroundColor: "themeBackground.brand", pt: ["3rem", "2rem"] }}>
+    <Box sx={{ backgroundColor: "themeBackground.brand", pt: ["3rem", "4rem", "2rem"] }}>
       <Container
         sx={{
           position: "relative",
         }}
       >
-        <GridBackground sx={{ pt: ["1rem", "2.6rem"] }}></GridBackground>
+        <GridBackground></GridBackground>
         <Announcement></Announcement>
+        <Box className={classes.vote}>
+          <Button color="primary" width={isMobile ? "22.4rem" : "28.5rem"} href={STICKER_CONTEST_DISCORD_VOTE_URL} target="_blank" whiteButton>
+            Vote on Discord Now
+          </Button>
+        </Box>
       </Container>
       <Award></Award>
     </Box>
@@ -62,10 +76,11 @@ const StickerContestResult = () => {
 
 const Announcement = () => {
   const { classes } = useStyles()
+  const { isPortrait } = useCheckViewport()
 
   return (
     <Stack direction="row" alignItems="center" gap={1} className={classes.announcement}>
-      <SvgIcon sx={{ fontSize: ["2.4rem", "4rem"] }} component={TadaSvg} inheritViewBox></SvgIcon>
+      <SvgIcon sx={{ fontSize: ["2.4rem", "4rem"] }} component={CalenderSvg} inheritViewBox></SvgIcon>
       <Typography
         sx={{
           fontSize: ["1.6rem", "3.2rem"],
@@ -75,7 +90,7 @@ const Announcement = () => {
           whiteSpace: "nowrap",
         }}
       >
-        The winners are here!!!
+        Voting ends:{isPortrait && <br></br>} Sun, Feb 11, 2024 23:59 UTC
       </Typography>
     </Stack>
   )
@@ -98,4 +113,4 @@ const Award = () => {
   )
 }
 
-export default StickerContestResult
+export default StickerContest
