+++ src/pages/sticker-vote/components/Gallery.tsx
