+++ src/pages/sticker-vote/components/GridBackground.tsx
