+++ src/pages/sticker-vote/components/ImageViewer.tsx
