+++ src/pages/sticker-vote/components/Title.tsx
