+++ src/pages/sticker-vote/index.tsx
