+++ src/pages/sticker-winners/Finalists/data.json
