+++ src/routes/index.tsx
@@ -18,8 +18,8 @@ import RollupScanBlock from "@/pages/rollup/block"
 import RollupScanChunk from "@/pages/rollup/chunk"
 import RollupScanChunkDetail from "@/pages/rollup/chunk/detail"
 import RollupScan from "@/pages/rollup/index"
-import StickerContest from "@/pages/sticker-contest"
-import StickerResult from "@/pages/sticker-contest-result"
+import StickerContest from "@/pages/sticker-vote"
+import StickerResult from "@/pages/sticker-winners"
 import Terms from "@/pages/terms"
 import { isSepolia } from "@/utils"
